Reorganize type checker files
