+++ ipe.cabal
@@ -36,6 +36,7 @@ library
       Ipe.Parser.TopLevelDefinition
       Ipe.Parser.TypeDefinition
       Ipe.TypeChecker
+      Ipe.TypeChecker.Expression
   other-modules:
       Paths_ipe
   hs-source-dirs:
@@ -74,7 +75,7 @@ test-suite ipe-test
       Ipe.Parser.TopLevelDefinitionSpec
       Ipe.Parser.TypeDefinitionSpec
       Ipe.ParserSpec
-      Ipe.TypeCheckerSpec
+      Ipe.TypeChecker.ExpressionSpec
       Paths_ipe
   hs-source-dirs:
       test
